Add linkerconfig to the host exports.
ART runs it on host to create the linker config for its chroot test
environment. This avoids problems building linkerconfig in the reduced
master-art manifest.

Even though linkerconfig currently is in platform/system I'm adding it
to the Bionic module exports based on discussions that the linkerconfig
binary should be part of the Runtime (Bionic) APEX due to its close
ties to the linker.

Test: build/soong/scripts/build-aml-prebuilts.sh runtime-module-host-exports
  then try the prebuilt snapshot with art/tools/buildbot-build.sh in
  master-art
Bug: 162819454
Change-Id: I4d43d7f0b980af48b193d5eb358e4f010eea541b