<arpa/nameser.h>: rationalize __INTRODUCED_IN.
Investigation shows that the symbols that claimed to have been in 32-bit
builds one API level earlier than in 64-bit builds actually weren't ---
everything was actually API level 22. This patch fixes the libc.map.txt
to match reality, and then simplifies the __INTRODUCED_IN()
incantations.

Investigation also shows that we have a bunch of unused #defines, so
this patch removes the ones that don't correspond to functions we
actually expose.

Test: treehugger
Change-Id: I540dd0d1d9561cac17c55eb68a07bed58dd718fa